@@ -1,12 +1,9 @@
-﻿using Microsoft.AspNetCore.Hosting;
-using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using Microsoft.Extensions.Hosting;
 
 namespace Lykke.MarginTrading.BrokerBase
 {
     internal static class Hosting
     {
-        internal static IWebHost WebHost { get; set; }
+        internal static IHost AppHost { get; set; }
     }
 }
